feat: test and sperate component for highlight
